refactor(Product): 엔티티로 수정
- 패키지 이동
- ProductRepository 추가
